@@ -9,54 +9,16 @@ module Spree
       @errors = ActiveModel::Errors.new(self)
     end
 
-    #
-    # Parameters can be passed using the following possible parameter configurations:
-    #
-    # * Single variant/quantity pairing
-    # +:variants => { variant_id => quantity }+
-    #
-    # * Multiple products at once
-    # +:products => { product_id => variant_id, product_id => variant_id }, :quantity => quantity+
-    def populate(from_hash)
+    def populate(variant_id, quantity, options = {}, price, spec)
 
-      # 17/10/13 DH: Store the dynamic BSC price for addition at 'OrderContents.add_to_line_item()' stage.
-      if from_hash[:price]
-        @order.contents.bscDynamicPrice = BigDecimal.new(from_hash[:price])
-      end
+      # 23/2/15 DH: Passing dynamic price and spec from web page
+      @order.contents.bscDynamicPrice = BigDecimal.new(price)
+      @order.contents.bscSpec = spec
     
-      # 22/11/13 DH: Now storing the BSC spec in the description part of the line item (so using same transfer mechanism as 'price')
-      #              Then we can add multiple curtains to the same order.
-      if from_hash[:spec]
-        @order.contents.bscSpec = from_hash[:spec]
-      end
-                
-      from_hash[:products].each do |product_id,variant_id|
-      
-        # 17/10/13 DH: If the hash contains a 'price' then set the variant's price
-        #              However this permanantly changes the variant's price so effects later use! 
-        #if from_hash[:price]
-        #  Spree::Variant.find(variant_id).default_price =  Spree::Price.new(:amount => from_hash[:price], :currency => currency)
-        #end
-      
-        attempt_cart_add(variant_id, from_hash[:quantity])
-      end if from_hash[:products]
-
-      from_hash[:variants].each do |variant_id, quantity|
-      
-        attempt_cart_add(variant_id, quantity)
-      end if from_hash[:variants]
-
-=begin
-      # 18/11/13 DH: Store the curtains spec in the 'Order.special_instructions'
-      #              Needs to be stored after 'attempt_cart_add' otherwise gets deleted for a new order.
-      if from_hash[:spec]
-        @order.special_instructions = from_hash[:spec]
-        # If the order isn't saved here then we loose the curtains spec (since 'special instructions' is normally used for 
-        # delivery info and doesn't get saved at this stage in the normal Spree Checkout state transition process)
-        @order.save!
-      end
-=end
-
+      ActiveSupport::Deprecation.warn "OrderPopulator is deprecated and will be removed from Spree 3, use OrderContents with order.contents.add instead.", caller
+      # protect against passing a nil hash being passed in
+      # due to an empty params[:options]
+      attempt_cart_add(variant_id, quantity, options || {})
       valid?
     end
 
@@ -66,32 +28,20 @@ module Spree
 
     private
 
-    def attempt_cart_add(variant_id, quantity)
+    def attempt_cart_add(variant_id, quantity, options = {})
       quantity = quantity.to_i
       # 2,147,483,647 is crazy.
       # See issue #2695.
       if quantity > 2_147_483_647
-        errors.add(:base, Spree.t(:please_enter_reasonable_quantity, :scope => :order_populator))
+        errors.add(:base, Spree.t(:please_enter_reasonable_quantity, scope: :order_populator))
         return false
       end
 
       variant = Spree::Variant.find(variant_id)
-      if quantity > 0
-
-        line_item = @order.contents.add(variant, quantity, currency)
-
-        bsc_req_id = 0
-        if (line_item.respond_to?(:bsc_req_id))
-          bsc_req_id = line_item.bsc_req_id
-        end
-        
-        #unless line_item.valid?
-        unless bsc_req_id > 0 && line_item.valid?
-          errors.add(:base, line_item.errors.messages.values.join(" "))
-          @order.contents.remove(variant)
-          return false
-        end
-      
+      begin
+        @order.contents.add(variant, quantity, options.merge(currency: currency))
+      rescue ActiveRecord::RecordInvalid => e
+        errors.add(:base, e.record.errors.messages.values.join(" "))
       end
     end
   end
